refactor(core): extraer helpers de tenants list, modules y HeaderEditor
- tenants/page.tsx: extraer PAGE_SIZE, tipos, labels, badges y formatDate
  a _helpers.tsx, reemplazar PHASE_BADGE_META inline con PhaseBadge (-128 líneas)
- modules/page.tsx: extraer ICON_NAMES catalog e IconPicker component
  a _helpers.tsx (-131 líneas)
- HeaderEditor.tsx: extraer tipos, NAV_LABELS, NAV_ICONS,
  INDUSTRY_HEADER_DEFAULTS y getField a header-editor-helpers.ts (-85 líneas)

Part of #244
